Add ml-example.py
Code for a generic machine learning example implemented as a datascience
pipeline (ingestion->wrangling->analysis->visualization).

The code takes a list of top level URLs as an input and does a first
level parsing of these web pages looking for CSV files.
The downloaded CSV files are then analysed as per three ML models viz.
LogisticRegression, RandomForestClassifier and GaussianNB.
Classification scores are stored in a dataframe and written as a csv
file at the end of the program.
A color coded heatmap is stored for a more visual comparison of the
models.
Visualization of the dataset in terms of scatter matrix, parallel
coordinates and radviz is also done and all generated images are stored
as png files.
The estimator i.e. model is stored on disk as a python pickle file.